feat: Refactor ThroughFocusSpotDiagram and update imports for clarity
