+++ optiland/analysis/__init__.py
@@ -1,6 +1,6 @@
 # flake8: noqa
 
-from .spot_diagram import SpotDiagram, ThroughFocusSpotDiagram
+from .spot_diagram import SpotDiagram
 from .encircled_energy import EncircledEnergy
 from .ray_fan import RayFan
 from .y_ybar import YYbar
@@ -10,18 +10,4 @@ from .field_curvature import FieldCurvature
 from .rms_vs_field import RmsSpotSizeVsField, RmsWavefrontErrorVsField
 from .pupil_aberration import PupilAberration
 from .irradiance import IncoherentIrradiance
-
-__all__ = [
-    "SpotDiagram",
-    "ThroughFocusSpotDiagram",
-    "EncircledEnergy",
-    "RayFan",
-    "YYbar",
-    "Distortion",
-    "GridDistortion",
-    "FieldCurvature",
-    "RmsSpotSizeVsField",
-    "RmsWavefrontErrorVsField",
-    "PupilAberration",
-    "IncoherentIrradiance",
-]
+from .through_focus_spot_diagram import ThroughFocusSpotDiagram
